davidson: fix edge case for complex matrices in
If the matrix is complex, we need phi to be complex so we can store
A*(thing) in it. If the norm of initial phi is small, we build a new
one; this new one also needs to be complex.

This fix guarantees that the new phi will have the same type as the
old phi.