Add movesites function
* Adds findsite[s](::MPS/MPO, is) for finding which sites of an MPS or MPO
contain specified indices.

* Add siteinds(::MPO) for getting the site indices of an MPO as a
vector of IndexSets.

* Add firstsiteinds(::MPO) for getting a vector of Index of the first
site indices of the MPO that are found.

* movesite(::MPS/MPO, n1 => m1) to move the site n1 to m1 by swapping
pairwise sites.

* movesites(::MPS/MPO, [n1 => m1, n2 => m2, ...]) to move the sites [n1,
n2, ...] to [m1, m2, ...] by swapping pairwise sites.

* MPS(A::ITensor, s::Vector{Index}; kwargs...) for decomposing an
ITensor A into an MPS, such that the site indices are [s[1], s[2], ...].
Truncation and dangling left and right link indices can be specified
with keyword arguments.

* MPO(A::ITensor, s::Vector{IndexSet}; kwargs...) for decomposing an
ITensor A into an MPO, such that the site indices are [s[1], s[2], ...].
Truncation and dangling left and right link indices can be specified
with keyword arguments.

* setindex!(psi::MPS/MPO, phi::MPS/MPO, n1:n2) for setting the MPS/MPO
tensors of psi between sites n1 and n2 to the tensors of the MPS/MPO
phi.

* setindex!(psi::MPS/MPO, A::ITensor, n1:n2; kwargs...) for
decomposing the ITensor A into an MPS/MPO and inserting it into the
sites n1:n2 of MPS/MPO psi. With aliases replacesites! and replacesites.

* Add swapbondsites[!] for MPO.