Making QNVal generic
The main intent here is to have the ability to explore what happens if you try to plug in QN which is not cocommutative, by replacing Int with something that has a noncommutative (+) method.

The prototypical example of that would be an MPO conserved quantity which is diagonal with regards to the Hilbert space indices, but which is noncommutative along the bond indices, along with a few slightly more exotic variations on that that can and do pop up as the most natural count of the number of excited particles for specific spin chains.

Hopefully the arrow & flux system will still work as long as the type has a bijective promotion to integers. Multiplication with arrows may require one to pick a small union of the desired type and its Int promotion for the flux computation.