Update Tutorials.md, correcting typo
In correcting this tiny typo, I would like to very sneakily ask if there is any plan to implement the Boson SiteType?? :angel:
I would love to see this feature implemented as it provides a plethora of new opportunities to study. Currently, I am a bit occupied but I would be interested in helping out implementing this feature!