Change `randomMPS` bond dim error to warning (#806)
Currently, `randomMPS` errors when the bond dimension doesn't reach
the desired one. This appears to happen fairly often with quantum
numbers, probably because of technicalities of quantum number
conservation (for example, to reach the desired bond dimension
it might be trying to add a block inconsistent with the flux).
This changes it into a more informative warning.