Change priority of `op` definition again
Change the ordering that `op` definitions get called to the following:

1. `op` with Index inputs that returns an ITensor.
2. `op` without Index inputs that returns a Matrix.
3. `op!` with Index inputs that modifies an ITensor in-place.

This makes it so that "Generic" definitions, which are defined using
the in-place version `op!`, get called last and don't take precedence
over `op` definitions if they exist.

Additionally, 1. gets placed in front of 2. to handle the case when
the `op` definition is defined with a Vararg of indices, to handle
the ambiguity when no indices are passed.