Add QR/RQ code to ensure all flux is moved onto R
When QR/RQ decomposing a zero flux tensor, this effectively ensures that there is zero flux between Q and R.  One can think of this a reduction in the gauge freedom between Q and R, making the results a little more deterministic.
Unit tests are also updated to check the zero flux was achieved.